These changes seem to be an update or modification of a JSON document parsing library. A few improvements have been made:
The `Node` class now has two methods: `IsDict()` and `AsDict()`, which replace the older `IsMap()` and `AsMap()` methods. This change implies that the internal representation for dictionaries (maps) might have changed in some way.

A new comparison operator `==` has been added to the `Node` class, which was not present before. This allows comparing two `Node` objects directly.

The constructor of the `Document` class now correctly initializes the member variable (`root_`). Before, it incorrectly initialized itself with a copy of the passed parameter, instead of moving it.
Also, some inline keywords were removed from function definitions, which can sometimes be beneficial for compiler optimization. However, they do not have a significant impact on functionality.

The `Print` function declaration was moved to after the definition of the `Document` class.

General formatting improvements, such as adding spaces inside of parens and braces, improving indentation, removing some unnecessary semicolons, and so on.

This code appears to be more C++11/14 compliant now, as it uses move semantics where appropriate, defines override specifiers for virtual functions (they're not shown in the provided snippet, though), and makes use of `auto` where possible.